refactor: move localStorage logic to camelCase service under localStorageService/
